Feat/apply testimonial freedomday to member

+++ src/main/kotlin/bass/repositories/DayRepository.kt
@@ -3,6 +3,6 @@ package bass.repositories
 import bass.entities.DayEntity
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface DayRepository : JpaRepository<DayEntity, Int> {
+interface DayRepository : JpaRepository<DayEntity, Long> {
     fun findByMemberId(memberId: Long): List<DayEntity>
 }
